[refactor] #133 UserDefaults 수정
- UserDefaults -> UserDefaultsClient 이름 수정
- Live, Test 분리
- 모듈 이름 주석 수정
