+++ Projects/CoreKit/Sources/Data/Client/UserDefaults/UserDefaultsClient+TestKey.swift
@@ -1,0 +1,14 @@
+//
+//  UserDefaultsClient+TestKey.swift
+//  CoreKit
+//
+//  Created by 김민호 on 9/30/24.
+//
+
+import Foundation
+
+import Dependencies
+
+extension UserDefaultsClient: TestDependencyKey {
+    public static var testValue = Self()
+}
